review: rename context.rs to verifier.rs
